+++ js/main.js
@@ -1,1 +1,2 @@
 import './data.js';
+import './create-picture.js';
