--- js/senteces.js
@@ -1,18 +1,0 @@
-import {getRandomInteger} from './random-number.js';
-
-const MESSAGETEXT = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
- // 1. Функция для выбора предложений из списка предложений
- const createArrayMessages = () => {
-  // Возвращаем элемент со случайным  номером выше из списка предложений
-  return MESSAGETEXT[getRandomInteger(0, MESSAGETEXT.length - 1)];
-};
-
-export{createArrayMessages};
