--- js/unique-sequential-number.js
@@ -1,10 +1,0 @@
-// Функция-генератор последовательных чисел
-function createIdGenerator () {
-  let lastGeneratedId = 0;
-  return function () {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-}
-
-export{createIdGenerator};
